rename/#77 : 컴포넌트 명과 파일명이 다른 문제를 해결한다
- 컴포넌트 명과 파일명을 일치시킨다.
