+++ src/app/(with-hamburger-button)/layout.tsx
@@ -1,4 +1,4 @@
-import ProtestList from '@/components/Protest/protest-list';
+import ProtestList from '@/components/Protest/ProtestList';
 import { ReactNode } from 'react';
 
 export default function Layout({ children }: { children: ReactNode }) {
