+++ src/components/Protest/MarkdownWrapper.tsx
@@ -3,11 +3,11 @@
 import React from 'react';
 import MarkdownPreview from '@uiw/react-markdown-preview';
 
-interface MarkdownWrapperProps {
+interface Props {
   content: string;
 }
 
-const MarkdownWrapper = ({ content }: MarkdownWrapperProps) => {
+export const MarkdownWrapper = ({ content }: Props) => {
   return (
     <div className='px-2 py-2 shadow-md mx-auto rounded-md w-[85%] min-w-[240px] flex items-center justify-center'>
       <MarkdownPreview
@@ -17,5 +17,3 @@ const MarkdownWrapper = ({ content }: MarkdownWrapperProps) => {
     </div>
   );
 };
-
-export default MarkdownWrapper;
