+++ src/components/Protest/ProtestDetailInfo.tsx
